+++ Magic/CHANGELOG.md
@@ -3,6 +3,7 @@
 # 7.8.1
 
  - Add invert_distance parameter to Damage action, to allow spells to do less damage up close.
+ - Add refund_on_no_target spell property, to refund costs on miss
 
 # 7.8